+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Pending
+# Fri Jul 24 21:41:51 UTC 2020
 
 * [Load V5] Fix Event Protocol and Cleanup [#16560](https://github.com/Applifier/unity-ads-webview/pull/16560)
 * [Load V5] Send load api enabled flag to kafka [#16434](https://github.com/Applifier/unity-ads-webview/pull/16434)