+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Pending
 
+* Replace adunitid macro for impression and operative tracking events [#16600](https://github.com/Applifier/unity-ads-webview/pull/16600)
 * [Load V5] Use Publisher configuration to enable Load V5 [#16435](https://github.com/Applifier/unity-ads-webview/pull/16435)
 
 # Mon Jul 27 23:16:58 UTC 2020