+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Pending
 
+* [Chore] Add tests to verify tracking urls in Load V5 [#16599](https://github.com/Applifier/unity-ads-webview/pull/16599)
 * [Load V5] Remove load v5 allow list from WebView [#16605](https://github.com/Applifier/unity-ads-webview/pull/16605)
 
 # Wed Jul 29 22:30:08 UTC 2020